tests/requirements.txt
playwright>=1.44
pytest
pytest-playwright
python-dotenv